backend/requirements.txt
# FastAPI and ASGI server
fastapi==0.115.6
uvicorn[standard]==0.34.0
mangum==0.19.0
gunicorn==23.0.0

# AWS SDK
boto3>=1.28.0
botocore>=1.31.0

# Database
psycopg[binary]>=3.1.0
SQLAlchemy>=2.0.20
alembic>=1.12.0

# Redis (ElastiCache)
redis[hiredis]>=5.0.01

# OpenSearch
opensearch-py==2.8.0
requests-aws4auth>=1.2.3

# Data validation
pydantic-settings==2.13.1

# HTTP client
httpx==0.28.1
aiohttp==3.11.12

# Web scraping
scrapy==2.12.0

# Kaggle API
kaggle==1.6.17

# Testing
pytest==8.3.4
pytest-asyncio==0.25.3
pytest-cov==6.1.0
hypothesis==6.130.6

# Utilities
python-dotenv==1.0.1
python-multipart==0.0.20
python-jose[cryptography]==3.3.0
